Move runtime/linux->runtime
1. Moves the linux specific runtime to be a generic runtime. This is the
first step in converging both linux/windows runtimes into a single
generic runtime based on a runc style differentiating layer.
