+++ cmd/containerd/builtins_linux.go
@@ -19,7 +19,7 @@ package main
 import (
 	_ "github.com/containerd/aufs"
 	_ "github.com/containerd/containerd/metrics/cgroups"
-	_ "github.com/containerd/containerd/runtime/linux"
+	_ "github.com/containerd/containerd/runtime"
 	_ "github.com/containerd/containerd/snapshots/native"
 	_ "github.com/containerd/containerd/snapshots/overlay"
 	_ "github.com/containerd/zfs"
